+++ app/src/main/java/kr/co/bootpay/android/SubscriptionBootpayPaymentActivity.java
@@ -46,7 +46,7 @@ public class SubscriptionBootpayPaymentActivity extends AppCompatActivity {
         items.add(item2);
 
         Payload payload = new Payload();
-        payload.setApplicationId("5b8f6a4d396fa665fdc2b5e8")
+        payload.setApplicationId(BootpayConstants.application_id)
                 .setOrderName("부트페이 결제테스트")
                 .setPg("나이스페이")
                 .setMethod("카드간편")